@@ -70,80 +70,85 @@
     "url": "https://github.com/bvaughn/react-virtualized/issues"
   },
   "devDependencies": {
-    "@babel/cli": "^7.7.0",
-    "@babel/core": "^7.7.2",
-    "@babel/plugin-external-helpers": "^7.2.0",
-    "@babel/plugin-proposal-class-properties": "^7.7.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.7.0",
-    "@babel/plugin-transform-runtime": "^7.6.2",
-    "@babel/plugin-transform-flow-comments": "^7.12.13",
-    "@babel/polyfill": "^7.7.0",
-    "@babel/preset-env": "^7.7.1",
-    "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-react": "^7.7.0",
-    "@babel/preset-stage-2": "^7.0.0",
-    "autoprefixer": "^9.7.1",
-    "babel-eslint": "^10.0.3",
-    "babel-jest": "^24.9.0",
-    "babel-loader": "8.0.6",
+    "@babel/cli": "^7.19.3",
+    "@babel/core": "^7.20.5",
+    "@babel/plugin-external-helpers": "^7.18.6",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-transform-flow-comments": "^7.19.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.19.6",
+    "@babel/plugin-transform-runtime": "^7.19.6",
+    "@babel/polyfill": "^7.12.1",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-flow": "^7.18.6",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-stage-2": "^7.8.3",
+    "@rollup/plugin-babel": "^6.0.3",
+    "@rollup/plugin-commonjs": "^23.0.3",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-replace": "^5.0.1",
+    "@rollup/plugin-terser": "^0.2.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/user-event": "^14.4.3",
+    "@uiw/react-codemirror": "^4.17.1",
+    "autoprefixer": "^10.4.13",
+    "babel-jest": "^29.3.1",
+    "babel-loader": "^9.1.0",
     "babel-plugin-flow-react-proptypes": "^26.0.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "codecov": "^3.6.1",
-    "codemirror": "^5.49.2",
-    "cross-env": "^6.0.3",
-    "css-loader": "^3.2.0",
-    "eslint": "^6.6.0",
-    "eslint-config-fbjs": "^3.1.1",
-    "eslint-config-prettier": "^6.5.0",
+    "codecov": "^3.8.3",
+    "codemirror": "^6.0.1",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.7.2",
+    "eslint": "^8.29.0",
+    "eslint-config-fbjs": "^4.0.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-config-react": "^1.1.7",
-    "eslint-plugin-babel": "^5.3.0",
-    "eslint-plugin-flowtype": "^4.3.0",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-relay": "^1.3.12",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^4.2.0",
-    "flow-bin": "^0.111.3",
-    "flow-copy-source": "^2.0.8",
-    "gh-pages": "^2.1.1",
-    "html-webpack-plugin": "^3.2.0",
-    "husky": "^3.0.9",
-    "immutable": "^4.0.0-rc.12",
-    "jest": "^24.9.0",
-    "jest-environment-puppeteer": "^4.3.0",
-    "lint-staged": "^9.4.2",
-    "postcss": "^7.0.21",
-    "postcss-cli": "^6.1.3",
-    "postcss-loader": "^3.0.0",
-    "prettier": "1.19.1",
-    "pretty-quick": "^2.0.1",
-    "puppeteer": "^2.0.0",
-    "react": "^17.0.1",
-    "react-codemirror": "^1.0.0",
-    "react-dom": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-test-renderer": "^17.0.1",
-    "rimraf": "^3.0.0",
-    "rollup": "^1.26.5",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-uglify": "^6.0.3",
-    "style-loader": "^1.0.0",
-    "watch": "^1.0.2",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0"
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-ft-flow": "^2.0.3",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-relay": "^1.8.3",
+    "eslint-visitor-keys": "^3.3.0",
+    "file-loader": "^6.2.0",
+    "flow-bin": "^0.194.0",
+    "flow-copy-source": "^2.0.9",
+    "gh-pages": "^4.0.0",
+    "hermes-eslint": "^0.9.0",
+    "html-webpack-plugin": "^5.5.0",
+    "husky": "^8.0.2",
+    "immutable": "^4.1.0",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "jest-environment-puppeteer": "^6.1.1",
+    "lint-staged": "^13.1.0",
+    "postcss": "^8.4.19",
+    "postcss-cli": "^10.1.0",
+    "postcss-loader": "^7.0.2",
+    "prettier": "^2.8.0",
+    "pretty-quick": "^3.1.3",
+    "puppeteer": "^19.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.4.4",
+    "react-router-dom": "^6.4.4",
+    "react-test-renderer": "^18.2.0",
+    "rimraf": "^3.0.2",
+    "rollup": "^3.6.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
-    "@babel/runtime": "^7.7.2",
-    "clsx": "^1.0.4",
-    "dom-helpers": "^5.1.3",
+    "@babel/runtime": "^7.20.6",
+    "@codemirror/lang-javascript": "^6.1.1",
+    "@uiw/codemirror-theme-darcula": "^4.17.1",
+    "clsx": "^1.2.1",
+    "dom-helpers": "^5.2.1",
     "loose-envify": "^1.4.0",
-    "prop-types": "^15.7.2",
+    "prop-types": "^15.8.1",
     "react-lifecycles-compat": "^3.0.4"
   },
   "peerDependencies": {
