@@ -1,8 +1,8 @@
-import nodeResolve from 'rollup-plugin-node-resolve';
-import commonjs from 'rollup-plugin-commonjs';
-import babel from 'rollup-plugin-babel';
-import replace from 'rollup-plugin-replace';
-import {uglify} from 'rollup-plugin-uglify';
+import resolve from '@rollup/plugin-node-resolve';
+import commonjs from '@rollup/plugin-commonjs';
+import babel from '@rollup/plugin-babel';
+import replace from '@rollup/plugin-replace';
+import terser from '@rollup/plugin-terser';
 
 export default {
   input: './source/index.js',
@@ -17,20 +17,24 @@ export default {
   },
   external: ['react', 'react-dom'],
   plugins: [
-    nodeResolve(),
+    resolve({
+      moduleDirectories: ['node_modules']
+    }),
     commonjs({
       include: 'node_modules/**',
     }),
     babel({
+      babelHelpers: 'external',
       exclude: 'node_modules/**',
     }),
     replace({
+      preventAssignment: true,
       'process.env.NODE_ENV': JSON.stringify('development'),
     }),
-    uglify({
+    terser({
       mangle: false,
       output: {
-        comments: true,
+        comments: "all",
         beautify: true,
       },
     }),
