@@ -31,20 +31,21 @@ export default function defaultRowRenderer({
     a11yProps.tabIndex = 0;
 
     if (onRowClick) {
-      a11yProps.onClick = event => onRowClick({event, index, rowData});
+      a11yProps.onClick = (event) => onRowClick({event, index, rowData});
     }
     if (onRowDoubleClick) {
-      a11yProps.onDoubleClick = event =>
+      a11yProps.onDoubleClick = (event) =>
         onRowDoubleClick({event, index, rowData});
     }
     if (onRowMouseOut) {
-      a11yProps.onMouseOut = event => onRowMouseOut({event, index, rowData});
+      a11yProps.onMouseOut = (event) => onRowMouseOut({event, index, rowData});
     }
     if (onRowMouseOver) {
-      a11yProps.onMouseOver = event => onRowMouseOver({event, index, rowData});
+      a11yProps.onMouseOver = (event) =>
+        onRowMouseOver({event, index, rowData});
     }
     if (onRowRightClick) {
-      a11yProps.onContextMenu = event =>
+      a11yProps.onContextMenu = (event) =>
         onRowRightClick({event, index, rowData});
     }
   }
